# src/dashboard/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "L2 Bot API",
    "description": "REST API для управления Lineage 2 бот-клиентом",
    "version": "1.0.0",
    "contact": {
      "name": "L2 Bot Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/connect": {
      "post": {
        "summary": "Connect to game server",
        "description": "Initiates connection to L2 game server",
        "operationId": "connect",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Connection initiated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/nearby/npcs": {
      "get": {
        "summary": "Get nearby NPCs",
        "description": "Returns list of NPCs/mobs in visibility range",
        "operationId": "getNearbyNpcs",
        "parameters": [
          {
            "name": "radius",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 600
            },
            "description": "Search radius"
          },
          {
            "name": "attackable",
            "in": "query",
            "schema": {
              "type": "boolean"
            },
            "description": "Only attackable NPCs"
          },
          {
            "name": "alive",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": true
            },
            "description": "Only alive NPCs"
          }
        ],
        "responses": {
          "200": {
            "description": "NPCs retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NearbyNpcsResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConnectRequest": {
        "type": "object",
        "properties": {
          "overrideConfig": {
            "type": "object",
            "properties": {
              "host": {
                "type": "string"
              },
              "loginPort": {
                "type": "integer"
              },
              "login": {
                "type": "string"
              },
              "password": {
                "type": "string"
              },
              "characterSlot": {
                "type": "integer"
              }
            }
          }
        }
      },
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "object"
          }
        }
      },
...
      "TargetResponse": {
        "type": "object",
        "properties": {
          "objectId": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": ["NPC", "PLAYER", "ITEM"]
          },
          "npcId": {
            "type": "integer"
          },
          "level": {
            "type": "integer"
          },
          "hp": {
            "type": "object",
            "properties": {
              "current": {
                "type": "integer"
              },
              "max": {
                "type": "integer"
              }
            }
          },
          "isAttackable": {
            "type": "boolean"
          },
          "isAggressive": {
            "type": "boolean"
          },
          "distance": {
            "type": "number"
          },
          "position": {
            "type": "object",
            "properties": {
              "x": {
                "type": "integer"
              },
              "y": {
                "type": "integer"
              },
              "z": {
                "type": "integer"
              }
            }
          }
        }
      },
...
      "NearbyNpcsResponse": {
        "type": "object",
        "properties": {
          "count": {
            "type": "integer"
          },
          "npcs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TargetResponse"
            }
          }
        }
      },
...
    }
  }
}